pyproject.toml
[project]
name = "speech-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "groq>=0.37.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
